Shoulder Slider Default
On startup the arm controls set the shoulder joint to the 0 position whereas the servos default to 90.  I changed the slider for the shoulder joint to reflect this discrepancy.  Turns out I had been setting the sliders' default position wrong all along and just got lucky with the default default.  That issue has now been corrected for all sliders.